+++ riverpie/lib/src/observer/history_observer.dart
@@ -1,0 +1,105 @@
+import 'package:riverpie/src/observer/event.dart';
+import 'package:riverpie/src/observer/observer.dart';
+
+/// The configuration of the [RiverpieHistoryObserver].
+/// It defines which events are saved.
+class HistoryObserverConfig {
+  /// Whether the observer should start immediately.
+  final bool startImmediately;
+
+  /// Whether the observer should save [ProviderInitEvent]s.
+  final bool saveProviderInitEvents;
+
+  /// Whether the observer should save [ListenerAddedEvent]s.
+  final bool saveListenerAddedEvents;
+
+  /// Whether the observer should save [ListenerRemovedEvent]s.
+  final bool saveListenerRemovedEvents;
+
+  /// Whether the observer should save [ChangeEvent]s.
+  final bool saveChangeEvents;
+
+  const HistoryObserverConfig({
+    this.startImmediately = true,
+    this.saveProviderInitEvents = false,
+    this.saveListenerAddedEvents = false,
+    this.saveListenerRemovedEvents = false,
+    this.saveChangeEvents = true,
+  });
+
+  /// By default, only [ChangeEvent]s are saved.
+  static const defaultConfig = HistoryObserverConfig();
+
+  /// Saves all events.
+  static const all = HistoryObserverConfig(
+    saveProviderInitEvents: true,
+    saveListenerAddedEvents: true,
+    saveListenerRemovedEvents: true,
+    saveChangeEvents: true,
+  );
+}
+
+/// An observer that stores every event in a list.
+/// This is useful for testing to keep track of the events.
+class RiverpieHistoryObserver extends RiverpieObserver {
+  /// The history of events.
+  final List<RiverpieEvent> history = [];
+
+  /// The configuration of the observer.
+  final HistoryObserverConfig config;
+
+  /// Whether the observer is currently listening to events.
+  bool listening;
+
+  RiverpieHistoryObserver([this.config = HistoryObserverConfig.defaultConfig])
+      : listening = config.startImmediately;
+
+  factory RiverpieHistoryObserver.all() {
+    return RiverpieHistoryObserver(HistoryObserverConfig.all);
+  }
+
+  @override
+  void handleEvent(RiverpieEvent event) {
+    if (!listening) {
+      return;
+    }
+
+    switch (event) {
+      case ProviderInitEvent _:
+        if (config.saveProviderInitEvents) {
+          history.add(event);
+        }
+        break;
+      case ListenerAddedEvent _:
+        if (config.saveListenerAddedEvents) {
+          history.add(event);
+        }
+        break;
+      case ListenerRemovedEvent _:
+        if (config.saveListenerRemovedEvents) {
+          history.add(event);
+        }
+        break;
+      case ChangeEvent _:
+        if (config.saveChangeEvents) {
+          history.add(event);
+        }
+        break;
+    }
+  }
+
+  /// Starts the observer to store events.
+  void start() {
+    listening = true;
+  }
+
+  /// Stops the observer from storing events.
+  void stop() {
+    listening = false;
+  }
+
+  /// Clears the history.
+  void clear() {
+    history.clear();
+  }
+}