+++ riverpie/test/provider/notifier_provider_test.dart
@@ -5,7 +5,7 @@ void main() {
   test('Single provider test', () {
     final notifier = _Counter(123);
     final provider = NotifierProvider<_Counter, int>((ref) => notifier);
-    final observer = RiverpieHistoryObserver();
+    final observer = RiverpieHistoryObserver.all();
     final ref = RiverpieContainer(
       observer: observer,
     );
@@ -45,7 +45,7 @@ void main() {
         providerB: ref.notifier(providerB),
       );
     });
-    final observer = RiverpieHistoryObserver();
+    final observer = RiverpieHistoryObserver.all();
     final ref = RiverpieContainer(
       observer: observer,
     );